Move Doctrine test fixtures under tests/fixtures
