+++ composer-dependency-analyser.php
@@ -13,7 +13,6 @@ $config
     ->addPathToScan(__DIR__ . '/migrations', false)
     ->addPathToScan(__DIR__ . '/src', false)
     ->addPathToScan(__DIR__ . '/web', false)
-    ->addPathToScan(__DIR__ . '/fixtures', true)
     ->addPathToScan(__DIR__ . '/tests', true)
     ->ignoreErrorsOnPackages([
         'doctrine/common',
